pokemons generals crud

+++ src/app/pokemons/favorites-pokemons/favorites-pokemons.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
 import { IFavorites, IPokemon } from 'src/app/services/interfaces';
 import { PokemonsServiceService } from 'src/app/services/pokemons-service.service';
+declare const $: any;
 
 @Component({
   selector: 'app-favorites-pokemons',
@@ -26,23 +27,51 @@ export class FavoritesPokemonsComponent implements OnInit {
 
   getFavorites(){
     this.favorites = this._pokemonService.getFavoritesList();
-      
   }
 
+  getFavorite(pokemon){
+    this.favoriteForm.patchValue(pokemon)
+    console.log(this.favoriteForm.value)
+  }
+
+  editFavoriteAlias(){
+    console.log(this.favoriteForm.value)
+    
+    if(!this.favoriteForm.valid){
+      return;
+    }
+ 
+    let response = this._pokemonService.editFavorite(this.favoriteForm.get('name').value, this.favoriteForm.get('alias').value);
+ 
+     if(response){
+       $('#editFavoritePokemon').modal('hide');
+       this.getFavorites();
+       alert('Pokemon Editado')
+ 
+     }else{
+       console.log('Pokemon no se pudo editar')
+     }
+  }
 
   deleteFavorite(item){
 
     this.favoriteForm.patchValue(item);
 
-    let response = this._pokemonService.deleteFromFavorites(this.favoriteForm.controls.name.value);
-
-    if(response){
-      console.log('Pokemon eliminado')
-      this.getFavorites();
-
-    }else{
-      console.log('Pokemon no se pudo eliminar')
+    if (confirm('Este pokemon se va a eliminar de tu lista de favoritos') == true) {      
+      let response = this._pokemonService.deleteFromFavorites(this.favoriteForm.controls.name.value);
+  
+      if(response){
+        this.getFavorites();
+        alert('Pokemon eliminado')
+  
+      }else{
+        console.log('Pokemon no se pudo eliminar')
+      }
     }
+    else{
+      return
+    }
+
   }
 
 }
